Add options for specifying bus and device
This adds the options --bus=<bus> and --device-number=<n>
for cases where there is more than one device with the same
vid:pid. In addition, -1 can be used as value if the device
is to be found on the last (highest) bus and/or device number.

These modifications are released into the public domain with
permission from my employer Host Mobility.